+++ cli/cli.go
@@ -28,10 +28,10 @@ import (
 	"github.com/essentialkaos/ek/v12/usage/update"
 
 	"github.com/essentialkaos/bibop/cli/executor"
+	"github.com/essentialkaos/bibop/cli/support"
 	"github.com/essentialkaos/bibop/parser"
 	"github.com/essentialkaos/bibop/recipe"
 	"github.com/essentialkaos/bibop/render"
-	"github.com/essentialkaos/bibop/support"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -115,16 +115,17 @@ func Init(gitRev string, gomod []byte) {
 	case options.Has(OPT_COMPLETION):
 		os.Exit(genCompletion())
 	case options.Has(OPT_GENERATE_MAN):
-		os.Exit(genMan())
+		genMan()
+		os.Exit(0)
 	case options.GetB(OPT_VER):
 		showAbout(gitRev)
-		return
+		os.Exit(0)
 	case options.GetB(OPT_VERB_VER):
-		support.ShowSupportInfo(APP, VER, gitRev, gomod)
-		return
+		support.Print(APP, VER, gitRev, gomod)
+		os.Exit(0)
 	case options.GetB(OPT_HELP) || len(args) == 0:
 		showUsage()
-		return
+		os.Exit(0)
 	}
 
 	configureSubsystems()
@@ -417,15 +418,13 @@ func genCompletion() int {
 }
 
 // genMan generates man page
-func genMan() int {
+func genMan() {
 	fmt.Println(
 		man.Generate(
 			genUsage(),
 			genAbout(""),
 		),
 	)
-
-	return 0
 }
 
 // genUsage generates usage info
