+++ cli/support/support.go
@@ -34,9 +34,9 @@ type Pkgs []Pkg
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// ShowSupportInfo prints verbose info about application, system, dependencies and
+// Print prints verbose info about application, system, dependencies and
 // important environment
-func ShowSupportInfo(app, ver, gitRev string, gomod []byte) {
+func Print(app, ver, gitRev string, gomod []byte) {
 	pkgs := collectEnvInfo()
 
 	fmtutil.SeparatorTitleColorTag = "{s-}"
